backend/requirements.txt

django==3.2
django-mysql==4.3
pymysql==1.0.2
cryptography==38.0.1  # required for local image build but not required on server... riiiight...
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
# mysqlclient==2.1.1
django-cors-headers==3.13.0
requests==2.28.1
gunicorn==20.1.0

